OpenVoxel: Significant work on world renderer

+++ OpenVoxel/src/main/java/net/openvoxel/common/block/IBlockAccess.java
@@ -15,9 +15,9 @@ public interface IBlockAccess {
 
 	boolean blockLoaded();
 
-	int getX();
-	int getY();
-	int getZ();
+	long getX();
+	long getY();
+	long getZ();
 
 	IBlockAccess getOffsetBlockData(BlockFace face);
